+++ src/components/GenericInlineForm/GenericInlineForm.jsx
@@ -282,6 +282,4 @@ GenericInlineForm.propTypes = {
 };
 
 
-export {
-  GenericInlineForm
-};
+export default GenericInlineForm;