+++ CHANGELOG.md
@@ -8,6 +8,15 @@ The big reworks
   - If you are using the `useAuth` hook and you are using the `refresh` endpoint of your api, you will need to update your api and create a new endpoint that will return a user's permissions.
     - The default endpoint is `/api/user/permissions`
 
+#### Components ####
+- `PamLayoutGrid` is now name `ConfigGrid`
+- `Button` has been removed. Use the MUI Button component.
+- `GenericForm` The predecessor to `ConfigForm` and `GenericConfigForm` has been removed. Use `ConfigForm` instead.
+
+#### Dead or Never used Components ####
+- `GenericLayout` has been removed.
+- `PamLayoutForm` has been removed. Use `ConfigForm` instead.
+
 ## Release 0.9.6 - 11/09/2023 ##
 Small release to fix a bug in the `viewLayout` hook.